Ultima aula

+++ 05-book-service/src/main/java/br/com/erudio/model/Book.java
@@ -32,7 +32,7 @@ public class Book implements Serializable {
 	@Temporal(TemporalType.DATE)
 	private Date launchDate;
 	@Column(nullable = false)
-	private BigDecimal price;
+	private Double price;
 	@Column(nullable = false, length = 250)
 	private String title;
 	@Transient
@@ -43,7 +43,7 @@ public class Book implements Serializable {
 	public Book() {
 	}
 
-	public Book(Long id, String author, String title, Date launchDate, BigDecimal price, String currency, String environment) {
+	public Book(Long id, String author, String title, Date launchDate, Double price, String currency, String environment) {
 		super();
 		this.id = id;
 		this.author = author;
@@ -78,11 +78,11 @@ public class Book implements Serializable {
 		this.launchDate = launchDate;
 	}
 
-	public BigDecimal getPrice() {
+	public Double getPrice() {
 		return price;
 	}
 
-	public void setPrice(BigDecimal price) {
+	public void setPrice(Double price) {
 		this.price = price;
 	}
